Move bitvec related parts into a separate file
