+++ src/lib.rs
@@ -58,6 +58,8 @@ pub mod alloc {
 mod codec;
 mod joiner;
 mod keyedvec;
+#[cfg(feature = "bit-vec")]
+mod bitvec;
 
 pub use self::codec::{Input, Output, Error, Encode, Decode, Codec, Compact, HasCompact, EncodeAsRef, CompactAs, EncodeAppend};
 pub use self::joiner::Joiner;
